middleware, pipe, validation, error catching, iterceptors, new endpoints for orders

+++ src/products/products.controller.ts
@@ -1,6 +1,17 @@
-import { Get, Param, Controller, Delete, Post, Body } from '@nestjs/common';
+import {
+  Get,
+  Param,
+  Controller,
+  Delete,
+  Post,
+  Body,
+  ParseUUIDPipe,
+  Put,
+  NotFoundException,
+} from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDTO } from './dtos/create-product.dto';
+import { UpdateProductDTO } from './dtos/update-product.dto';
 
 @Controller('products')
 export class ProductsController {
@@ -9,8 +20,10 @@ export class ProductsController {
   }
 
   @Get('/:id')
-  public getById(@Param('id') id: string) {
-    return this.productsService.getById(id);
+  getById(@Param('id', new ParseUUIDPipe()) id: string) {
+    const prod = this.productsService.getById(id);
+    if (!prod) throw new NotFoundException('Product not found');
+    return prod;
   }
 
   @Get('/')
@@ -19,7 +32,9 @@ export class ProductsController {
   }
 
   @Delete('/:id')
-  deleteById(@Param('id') id: string) {
+  deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
     this.productsService.deleteById(id);
     return { success: true };
   }
@@ -27,5 +42,17 @@ export class ProductsController {
   @Post('/')
   create(@Body() productData: CreateProductDTO) {
     return this.productsService.create(productData);
+  }
+
+  @Put('/:id')
+  update(
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() productData: UpdateProductDTO,
+  ) {
+    if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
+
+    this.productsService.updateById(id, productData);
+    return { success: true };
   }
 }
